radix sort: make the fused kernel the sole device_radix_sort; remove multi-launch
Fold _radix_sort_fused.py back into _radix_sort.py so the single fused kernel
*is* device_radix_sort. The merged driver keeps the old signature (values,
tmp_values, end_bit, scratch=None) plus the device-N graph extras (n,
log256_max_n), and retains the shared-scratch fallback when scratch is None so
the existing scratch-budget/error-path tests keep passing (removing _scratch.py
stays a separate follow-up).

- Relocate RADIX_BITS/RADIX_DIGITS/InsufficientScratchError into _radix_sort.py;
  delete the old multi-launch implementation and its histogram/scan/scatter
  kernels.
- Add host device_radix_sort_scratch_slots(n) (default depth, 0 for n<=1) on top
  of the kernel-callable fused_radix_sort_scratch_slots(n, log256_max_n); the
  forced-depth footprint equals the old recursive scan footprint.
- Drop the redundant Stage-1 fused tests (dtype x size x value now exercised by
  test_device_radix_sort_keys_only/_key_value on the fused kernel); keep the
  explicit-depth test.
- test_api: add InsufficientScratchError + the two scratch-slots helpers to the
  expected qd.algorithms surface; drop device_radix_sort_fused.
