+++ python/quadrants/algorithms/__init__.py
@@ -5,9 +5,6 @@ from ._radix_sort import (
     InsufficientScratchError,
     device_radix_sort,
     device_radix_sort_scratch_slots,
-)
-from ._radix_sort_fused import (
-    device_radix_sort_fused,
     fused_radix_sort_scratch_slots,
 )
 from ._reduce import device_reduce_add, device_reduce_max, device_reduce_min
@@ -26,7 +23,6 @@ __all__ = [
     "device_exclusive_scan_max",
     "device_exclusive_scan_min",
     "device_radix_sort",
-    "device_radix_sort_fused",
     "device_radix_sort_scratch_slots",
     "device_reduce_add",
     "device_reduce_by_key_add",
